synapse_flask/requirements.txt
Flask>=3.0,<4.0
Flask-SocketIO>=5.5,<6.0
psycopg[binary]>=3.2,<4.0
paho-mqtt>=2.0,<3.0
python-dotenv>=1.0,<2.0
simple-websocket>=1.1,<2.0
